feat(ffe-accordion): Decrease need for modifier classes
BREAKING CHANGE: The modified class `ffe-accordion--blue` supersedes
the modified class `ffe-accordion-item--blue` and all element level
classes of `ffe-accordion-item` with the `--blue` modifier, i.e.
`ffe-accordion-item__(toggler,icon,content)--blue` are no longer in use.

The modified class `ffe-accordion-item--open` is still needed, but the
modifier `--open` on element level classes in this block are not
necessary anymore.